[Data] Move resource budget Prometheus gauges to ExecutionCallback (#62209)
### Description

This PR moves out responsibility of tracking Prometheus metrics for
resource budgets into a new class ResourceAllocatorPrometheusCallback.
This is in accordance with single responsibility principle.

Closes #60269

---------
