+++ python/ray/data/_internal/execution/callbacks/__init__.py
@@ -1,7 +1,11 @@
 from ray.data._internal.execution.callbacks.insert_issue_detectors import (
     IssueDetectionExecutionCallback,
 )
+from ray.data._internal.execution.callbacks.resource_allocator_prometheus_callback import (
+    ResourceAllocatorPrometheusCallback,
+)
 
 __all__ = [
     "IssueDetectionExecutionCallback",
+    "ResourceAllocatorPrometheusCallback",
 ]
