+++ python/ray/data/context.py
@@ -984,11 +984,15 @@ class DataContext:
         from ray.data._internal.execution.callbacks.insert_issue_detectors import (
             IssueDetectionExecutionCallback,
         )
+        from ray.data._internal.execution.callbacks.resource_allocator_prometheus_callback import (
+            ResourceAllocatorPrometheusCallback,
+        )
         from ray.data._internal.execution.execution_callback import ExecutionCallback
 
         classes = [
             ExecutionIdxUpdateCallback,
             IssueDetectionExecutionCallback,
+            ResourceAllocatorPrometheusCallback,
         ]
 
         # Parse environment variable for custom callbacks
